Move statistics to WatchingVirtualFileSystem
We don't need the statistics anywhere else.
