+++ subprojects/core/src/main/java/org/gradle/internal/vfs/RoutingVirtualFileSystem.java
@@ -95,11 +95,6 @@ public class RoutingVirtualFileSystem implements VirtualFileSystem {
         getVirtualFileSystemFor(snapshot.getAbsolutePath()).updateWithKnownSnapshot(snapshot);
     }
 
-    @Override
-    public VirtualFileSystemStatistics getStatistics() {
-        return gradleUserHomeVirtualFileSystem.getStatistics();
-    }
-
     private VirtualFileSystem getVirtualFileSystemFor(String location) {
         return vfsRetained.getAsBoolean()
             ? gradleUserHomeVirtualFileSystem
